move components to containers that access redux
