+++ src/client/src/containers/SelectFrontEndFramework/index.tsx
@@ -1,9 +1,7 @@
 import * as React from "react";
 import { connect } from "react-redux";
 
-import SelectBackendFramework from "../SelectBackendFramework";
-
-import SelectOption from "../SelectOption";
+import SelectOption from "../../components/SelectOption";
 
 import { getFrontendFrameworksAction } from "../../actions/getFrontendFrameworks";
 import { selectFrontendFramework as selectFrontendAction } from "../../actions/selectFrontEndFramework";
@@ -46,7 +44,6 @@ class SelectFrontEndFramework extends React.Component<Props> {
           options={this.props.options}
           selectedCards={this.convertSelectionToIndexNumber(this.props.selectedFramework)}
         />
-        <SelectBackendFramework />
       </div>
     );
   }
